+++ activejob/test/support/integration/adapters/sidekiq.rb
@@ -49,7 +49,7 @@ module SidekiqJobsManager
   def can_run?
     begin
       Sidekiq.redis { |conn| conn.connect }
-    rescue => e
+    rescue
       return false
     end
     true